xorgxserver: work with xinit
If xinit is to be run by unprivileged users, the Xorg server must be
setuid.  However, Xorg accepts limited command line options when run
with elevated privileges.  So build working defaults into the program.

Further, the code in xorgserver-xkbcomp-path.patch to access XKB_BINDIR
is not executed in the absence of Xkb-related command line options (one
of which, -xkbdir, is disallowed when Xorg is setuid).  Move that code
to a place where it will be executed.